+++ cmd/cli/test/sharding_on_test.go
@@ -1,0 +1,9 @@
+//go:build sharding
+
+// Package test provides E2E tests of AIS CLI
+/*
+ * Copyright (c) 2018-2025, NVIDIA CORPORATION. All rights reserved.
+ */
+package test_test
+
+const shardingEnabled = true