+++ ext/dsort/bcast.go
@@ -1,3 +1,5 @@
+//go:build sharding
+
 // Package dsort provides distributed massively parallel resharding for very large datasets.
 /*
  * Copyright (c) 2018-2025, NVIDIA CORPORATION. All rights reserved.